fix formData for uploading img
