+++ api/client.js
@@ -10,6 +10,7 @@ export const getMyURL = "https://orbital-funfit.herokuapp.com/routine/getMyRouti
 export const deleteRoutineURL = "https://orbital-funfit.herokuapp.com/routine/deleteRoutine";
 export const editRoutineURL = "https://orbital-funfit.herokuapp.com/routine/editRoutine";
 export const uploadImageURL = "https://orbital-funfit.herokuapp.com/user/upload";
+export const downloadPicURL = "https://orbital-funfit.herokuapp.com/user/downloadPic";
 
 //export default axios.create({baseURL: "https://orbital-funfit.herokuapp.com"});
 // export const loginURL = "https://secret-garden-61740.herokuapp.com/https://orbital-funfit.herokuapp.com/user/login";
